Move AddTracing to tracing.go
